development/nephrology-ma/pyproject.toml
[project]
name = "nephrology-ma"
version = "0.1.0"
description = "Nephrology M&A Intelligence Platform — acquisition target scoring from free CMS data"
readme = "AGENTS.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
    { name = "The Kidney Experts", email = "shree.mulay@thekidneyexperts.com" },
]

dependencies = [
    "google-cloud-bigquery>=3.25",
    "google-cloud-bigquery-storage>=2.25",
    "google-auth>=2.29",
    "gspread>=6.1",
    "google-auth-oauthlib>=1.2",
    "pandas>=2.2",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "pyarrow>=16.0",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "db-dtypes>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
